Labels

+++ src/main/java/swp/group2/learninghub/service/CoreLabelsService.java
@@ -4,8 +4,8 @@ import swp.group2.learninghub.model.CoreLabel;
 import java.util.List;
 public interface CoreLabelsService {
     List<CoreLabel> getAllLabels();
-    CoreLabel getLabelById(Long id);
+    CoreLabel getLabelById(int id);
     CoreLabel createLabel(CoreLabel label);
     CoreLabel updateLabel(CoreLabel label);
-    void deleteLabel(Long id);
+    void deleteLabel(int id);
 }
